requirements.txt
# ============================================================
# Lead Generation & Outreach System - Python Dependencies
# ============================================================
# Install with:  pip install -r requirements.txt

# Module 1 - Prospect Database
# (SQLite is built into Python - no driver needed)

# Pretty-print tables in the terminal
tabulate==0.9.0

# Module 2 - Lead Scorer (rule-based, no API key required)

# Module 7 - SMTP Mailer
python-dotenv>=1.0.0

# Module 9 - AI Engine
anthropic>=0.49.0

# Module 10 - Google Maps Scraper
googlemaps>=4.10.0

# Module 11 - Web Dashboard
Flask>=3.0.0

# Module 12 - Deep Research Agent
beautifulsoup4>=4.12.0
requests>=2.31.0

# Module 13 - PDF Proposal Generator
reportlab>=4.0.0
python-pptx>=0.6.23

# Module 14 - Social Automation
playwright>=1.40.0

# Module 15 - Twilio SMS
twilio>=8.0.0

# Module 16 - SendGrid Infrastructure
sendgrid>=6.11.0
